First code commit and initial prototype
Finally got this to a point where there's enough of a direction that
it feels like this is ready for a commit. Working prototype-level
support for type signatures and type synonyms, with the quasiquoter in
both the declaration and type context. The latter is necessary for
locals, as quasiquoters are only allowed to generate top-level
declarations. In fact, a lot of the "fancy" stuff in here (and there's a
lot of it) exists to support local signatures and work around these
limitations.

The giant, kludgy partial copy of Fixpoint types is only temporary,
until I can work out how to properly store the InternedText that the
Fixpoint AST wants. Types will be broken out into separate packages
eventually, so that's going to necessitate some changing around of
things anyway.