Split into multiple modules
liquidhaskell-core contains the core types and functions that don't have
a GHC dependency.

liquidhaskell-lang is the module to be depended on by
LiquidHaskell-using projects. It has minimal dependencies (as does
liquidhaskell-core), and implements an `lq` quasiquoter which simply
spits out plain Haskell types without refinements.

liquidhaskell-ghc contains the types and functions that depend on GHC
and can be reused elsewhere.

liquidhaskell-frontend is the main "liquidhaskell" program. It also
includes a module Language.Haskell.Liquid.Quasi which replaces the
LiquidHaskell module from liquidhaskell-lang during compilation with the
frontend. This module implements the full annotation generation and
such, but depends on GHC. This is why we have the liquidhaskell-lang
module: so LiquidHaskell-using projects can be compiled normally without
a GHC dependency (or any of the other various dependencies that the rest
of LH pulls in).

liquidhaskell-frontend will eventually become one with the liquidhaskell
package.