Parser captures both tycons and datacons in exprs
Instead of assuming a constructor referenced in an Expr is a promoted
data constructor, capture either a tycon or a datacon and let GHC figure
out which it should be.

Later, we'll have the reifier detect which one GHC resolved it to and
try to expand it as an expression alias if it's a tycon. We'll raise an
error if the alias application doesn't come out to kind Expr (i.e. the
alias isn't an expression alias). This is where the location information
on ECtr will come in handy.